feat: expose compiler as part of dbc package
This commit moves the DBC compiler to the `pkg/dbc` package to make it
part of the public API.

Currently, if you want to do anything useful with a DBC file you more
or less need to re-implement parts of the compiler yourself, which
gets tedious to keep up to date.
